This adds a mechanism for "publishing" internal APIs for use across our libraries.
A client library loads the API class "BD_InternalAPI_Capture" and then calls "exposeAPI" on it to link up at runtime.
If the class fails to load, or the requested API doesn't exist, it downgrades gracefully to a no-op rather than crashing.

See InternalAPI.h for a more thorough description + example client.