perform full LTO during Android .so build using rustc (#649)
Previously we were building a cc_binary target which seemed to have an issue with the version of LLVM used that prevented LTO from working. Based on some comments it meant that we were not enabling LTO for this reason (despite flags that might indicate this).

Realizing that cargo was able to do this, by moving the final link to rust_shared_library we can use rustc for everything which simplified this and allows us to run fat LTO as part of the .so build.

To get this working I needed to update the std rlib to be built with LTO compile flags.

In addition, this brings in Rust 1.90.0 for both Android and iOS.