Remove llvm pinning (#695)
Including llvm as a toolchain is great for creating hermetic builds but it slows down our pipeline a lot. The tar is multi GB and takes minutes to decompress.

The chances of a different llvm-ar version to change the output is slim to none.

Note to future self: If this becomes an issue we can at least use llvm-ar from the llvm-tools pulled from rust